+++ admin/src/scenes/phase0/commons.ts
@@ -4,6 +4,12 @@ import * as fr from "i18n-iso-countries/langs/fr.json";
 countries.registerLocale(fr);
 const countriesList = countries.getNames("fr", { select: "official" });
 
+export const REJECTION_REASONS = {
+  NOT_FRENCH: "Le volontaire n'est pas de nationalité française",
+  TOO_YOUNG: "Le volontaire n'a pas l'âge requis",
+  OTHER: "Autre (préciser)",
+};
+
 export const SPECIFIC_SITUATIONS_KEY = [
   "qpv",
   "handicap",
